+++ src/components/niko-table/config/index.ts
