+++ src/components/niko-table/core/index.ts
