+++ src/content/docs/examples/row-expansion-table.mdx
@@ -67,11 +67,9 @@ First, we'll define our columns with an expand column.
 ```tsx showLineNumbers title="columns.tsx"
 "use client"
 
-import {
-  DataTableColumnHeader,
-  DataTableColumnTitle,
-  DataTableColumnSortMenu,
-} from "@/components/niko-table/components"
+import { DataTableColumnHeader } from "@/components/niko-table/components/data-table-column-header"
+import { DataTableColumnTitle } from "@/components/niko-table/components/data-table-column-title"
+import { DataTableColumnSortMenu } from "@/components/niko-table/components/data-table-column-sort"
 import type { DataTableColumnDef } from "@/components/niko-table/types"
 import { Button } from "@/components/ui/button"
 import { ChevronRight, ChevronDown } from "lucide-react"
@@ -189,15 +187,13 @@ import {
   DataTableBody,
   DataTableEmptyBody,
 } from "@/components/niko-table/core"
-import {
-  DataTableToolbarSection,
-  DataTablePagination,
-  DataTableSearchFilter,
-  DataTableViewMenu,
-  DataTableColumnHeader,
-  DataTableColumnTitle,
-  DataTableColumnSortMenu,
-} from "@/components/niko-table/components"
+import { DataTableToolbarSection } from "@/components/niko-table/components/data-table-toolbar-section"
+import { DataTablePagination } from "@/components/niko-table/components/data-table-pagination"
+import { DataTableSearchFilter } from "@/components/niko-table/components/data-table-search-filter"
+import { DataTableViewMenu } from "@/components/niko-table/components/data-table-view-menu"
+import { DataTableColumnHeader } from "@/components/niko-table/components/data-table-column-header"
+import { DataTableColumnTitle } from "@/components/niko-table/components/data-table-column-title"
+import { DataTableColumnSortMenu } from "@/components/niko-table/components/data-table-column-sort"
 import type { DataTableColumnDef } from "@/components/niko-table/types"
 import { Button } from "@/components/ui/button"
 import { ChevronRight, ChevronDown } from "lucide-react"
@@ -413,7 +409,7 @@ You can export table data to CSV format. When exporting row expansion tables, ke
 The simplest way to add export functionality:
 
 ```tsx showLineNumbers
-import { DataTableExportButton } from "@/components/niko-table/components"
+import { DataTableExportButton } from "@/components/niko-table/components/data-table-export-button"
 
 export function RowExpansionTableWithExport({ data }: { data: Order[] }) {
   return (
@@ -452,7 +448,7 @@ For more control, use the `exportTableToCSV` function directly:
 
 ```tsx showLineNumbers
 import { useDataTable } from "@/components/niko-table/core"
-import { exportTableToCSV } from "@/components/niko-table/filters"
+import { exportTableToCSV } from "@/components/niko-table/filters/table-export-button"
 import { Button } from "@/components/ui/button"
 import { Download } from "lucide-react"
