+++ src/content/docs/examples/row-selection-table.mdx
@@ -67,11 +67,9 @@ First, we'll add a select column to our definitions.
 ```tsx showLineNumbers title="columns.tsx"
 "use client"
 
-import {
-  DataTableColumnHeader,
-  DataTableColumnTitle,
-  DataTableColumnSortMenu,
-} from "@/components/niko-table/components"
+import { DataTableColumnHeader } from "@/components/niko-table/components/data-table-column-header"
+import { DataTableColumnTitle } from "@/components/niko-table/components/data-table-column-title"
+import { DataTableColumnSortMenu } from "@/components/niko-table/components/data-table-column-sort"
 import type { DataTableColumnDef } from "@/components/niko-table/types"
 import { Checkbox } from "@/components/ui/checkbox"
 
@@ -164,15 +162,13 @@ import {
   DataTableBody,
   DataTableEmptyBody,
 } from "@/components/niko-table/core"
-import {
-  DataTableToolbarSection,
-  DataTablePagination,
-  DataTableSearchFilter,
-  DataTableViewMenu,
-  DataTableColumnHeader,
-  DataTableColumnTitle,
-  DataTableColumnSortMenu,
-} from "@/components/niko-table/components"
+import { DataTableToolbarSection } from "@/components/niko-table/components/data-table-toolbar-section"
+import { DataTablePagination } from "@/components/niko-table/components/data-table-pagination"
+import { DataTableSearchFilter } from "@/components/niko-table/components/data-table-search-filter"
+import { DataTableViewMenu } from "@/components/niko-table/components/data-table-view-menu"
+import { DataTableColumnHeader } from "@/components/niko-table/components/data-table-column-header"
+import { DataTableColumnTitle } from "@/components/niko-table/components/data-table-column-title"
+import { DataTableColumnSortMenu } from "@/components/niko-table/components/data-table-column-sort"
 import type { DataTableColumnDef } from "@/components/niko-table/types"
 import { Checkbox } from "@/components/ui/checkbox"
 import { Button } from "@/components/ui/button"
@@ -394,7 +390,7 @@ function BulkActions() {
 Use `DataTableSelectionBar` to show a persistent selection bar:
 
 ```tsx showLineNumbers
-import { DataTableSelectionBar } from "@/components/niko-table/components"
+import { DataTableSelectionBar } from "@/components/niko-table/components/data-table-selection-bar"
 
 <DataTableSelectionBar
   selectedCount={selectedRows.length}
