+++ src/registry/new-york/examples/niko-table/virtualized-column-dnd.tsx
@@ -7,13 +7,13 @@ import {
   DataTableVirtualizedDndHeader,
   DataTableVirtualizedDndColumnBody,
 } from "@/components/niko-table/core/data-table-virtualized-dnd-structure"
+import { DataTableColumnDndProvider } from "@/components/niko-table/components/data-table-column-dnd"
 import {
-  DataTableColumnDndProvider,
   DataTableEmptyIcon,
   DataTableEmptyMessage,
   DataTableEmptyTitle,
   DataTableEmptyDescription,
-} from "@/components/niko-table/components"
+} from "@/components/niko-table/components/data-table-empty-state"
 import type { DataTableColumnDef } from "@/components/niko-table/types"
 import { Badge } from "@/components/ui/badge"
 import { Inbox } from "lucide-react"
